+++ wetterdienst/api.py
@@ -5,6 +5,7 @@
 import importlib
 from typing import ClassVar
 
+from wetterdienst.exceptions import ApiNotFoundError
 from wetterdienst.model.request import TimeseriesRequest
 
 
@@ -67,17 +68,17 @@ class Wetterdienst:
         try:
             module_path, class_name = cls.registry[provider][network].rsplit(".", 1)
         except KeyError as e:
-            msg = f"No API available for provider {provider} and network {network}"
-            raise KeyError(msg) from e
+            msg = f"No API available for provider {provider} and network {network}."
+            raise ApiNotFoundError(msg) from e
 
         try:
             module = importlib.import_module(module_path)
             return getattr(module, class_name)
         except ModuleNotFoundError as e:
-            msg = f"Module {module_path} not found"
+            msg = f"Module {module_path} not found."
             raise ImportError(msg) from e
         except AttributeError as e:
-            msg = f"Class {class_name} not found in module {module_path}"
+            msg = f"Class {class_name} not found in module {module_path}."
             raise AttributeError(msg) from e
 
     def __new__(cls, provider: str, network: str) -> type[TimeseriesRequest]: